requirements.txt
Flask
Flask-SQLAlchemy
PyYAML
pdfplumber
agno
Werkzeug
gunicorn         # If deploying to production (optional)